pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-learning-agents"
version = "0.1.0"
description = "Agentic AI framework: multi-agent orchestration + BQ Terraform generation."
requires-python = ">=3.10"
dependencies = [
  # Cloud LLM backends
  "google-cloud-aiplatform>=1.60.0",
  # Offline LLM backend (Ollama REST client)
  "requests>=2.31.0",
  # PDF schema extraction
  "pdfplumber>=0.10.0",
  # Environment config
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
run-agents       = "ai_agents.scripts.run_agents:main"
run-bq-terraform = "ai_agents.scripts.run_bq_terraform:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
